requirements/base.txt
#
# This file is autogenerated by pip-compile with python 3.8
# To update, run:
#
#    pip-compile requirements/base.in
#
antlr4-python3-runtime==4.9.3
    # via omegaconf
certifi==2022.9.24
    # via requests
cffi==1.15.1
    # via cryptography
charset-normalizer==2.1.1
    # via
    #   pdfminer-six
    #   requests
contourpy==1.0.5
    # via matplotlib
cryptography==38.0.1
    # via pdfminer-six
cycler==0.11.0
    # via matplotlib
effdet==0.3.0
    # via layoutparser
filelock==3.8.0
    # via huggingface-hub
fonttools==4.37.4
    # via matplotlib
huggingface-hub==0.10.1
    # via timm
idna==3.4
    # via requests
iopath==0.1.10
    # via layoutparser
kiwisolver==1.4.4
    # via matplotlib
layoutparser[layoutmodels,tesseract]==0.3.4
    # via -r requirements/base.in
matplotlib==3.6.1
    # via pycocotools
numpy==1.23.4
    # via
    #   contourpy
    #   layoutparser
    #   matplotlib
    #   opencv-python
    #   pandas
    #   pycocotools
    #   scipy
    #   torchvision
omegaconf==2.2.3
    # via effdet
opencv-python==4.6.0.66
    # via layoutparser
packaging==21.3
    # via
    #   huggingface-hub
    #   matplotlib
    #   pytesseract
pandas==1.5.0
    # via layoutparser
pdf2image==1.16.0
    # via layoutparser
pdfminer-six==20220524
    # via pdfplumber
pdfplumber==0.7.5
    # via layoutparser
pillow==9.2.0
    # via
    #   layoutparser
    #   matplotlib
    #   pdf2image
    #   pdfplumber
    #   pytesseract
    #   torchvision
portalocker==2.5.1
    # via iopath
pycocotools==2.0.5
    # via effdet
pycparser==2.21
    # via cffi
pyparsing==3.0.9
    # via
    #   matplotlib
    #   packaging
pytesseract==0.3.10
    # via layoutparser
python-dateutil==2.8.2
    # via
    #   matplotlib
    #   pandas
pytz==2022.5
    # via pandas
pyyaml==6.0
    # via
    #   huggingface-hub
    #   layoutparser
    #   omegaconf
    #   timm
requests==2.28.1
    # via
    #   huggingface-hub
    #   torchvision
scipy==1.9.2
    # via layoutparser
six==1.16.0
    # via python-dateutil
timm==0.6.11
    # via effdet
torch==1.12.1
    # via
    #   effdet
    #   layoutparser
    #   timm
    #   torchvision
torchvision==0.13.1
    # via
    #   effdet
    #   layoutparser
    #   timm
tqdm==4.64.1
    # via
    #   huggingface-hub
    #   iopath
typing-extensions==4.4.0
    # via
    #   huggingface-hub
    #   iopath
    #   torch
    #   torchvision
urllib3==1.26.12
    # via requests
wand==0.6.10
    # via pdfplumber
